intellicrack/requirements.txt
aiohttp==3.12.13
angr==9.2.158
anthropic==0.54.0
azure_storage==0.37.0
boto3==1.38.41
cairocffi==1.7.1
capstone==5.0.3
celery==5.5.3
cffi==1.17.1
claripy==9.2.158
click==8.1.7
coverage==7.8.2
cryptography==45.0.4
cupy==13.4.1
dask==2025.5.1
dpkt_fix==1.7
Flask==3.1.1
flask_cors==6.0.1
frida==17.2.4
intel_extension_for_pytorch==2.7.0
Jinja2==3.1.6
joblib==1.5.1
keystone_engine==0.9.2
langchain==0.3.26
lief==0.16.6
macholib==1.16.3
manticore==0.3.7
matplotlib==3.10.3
networkx==3.1
nltk==3.9.1
numba==0.61.2
numpy==2.3.1
onnx==1.18.0
onnxruntime==1.22.0
openai==1.90.0
pandas==2.3.0
paramiko==3.5.1
# pcap==1.1  # Skipped - using scapy instead for packet capture
# pcapy==0.11.4  # Skipped - using scapy instead for packet capture
pdfkit==1.0.0
pefile==2024.8.26
peft==0.15.2
Pillow==11.2.1
psutil==7.0.0
psycopg2_binary==2.9.10
pycuda==2025.1.1
pyelftools==0.32
pymongo==4.13.2
pynput==1.8.1
pyopencl==2025.2.4
pyOpenSSL==25.1.0
pypykatz==0.6.11
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyQt6==6.9.1
pyshark==0.6
python-dotenv==1.1.0
pyusb==1.3.1
pywin32==310
PyYAML==6.0.2
PyYAML==6.0.2
qiling==1.4.6
r2pipe==1.9.4
radon==6.0.1
ray==2.47.1
reportlab==4.4.1
Requests==2.32.4
rich==14.0.0
scapy==2.6.1
scikit_learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
tensorflow_cpu==2.19.0
tinycss2==1.4.0
torch==2.7.0
transformers==4.52.4
triton==3.3.1
watchdog==6.0.0
weasyprint==65.1
wmi==1.5.1
xattr==1.1.4
xlsxwriter==3.2.5
z3_solver==4.13.0.0
